+++ src/main/java/me/ultrusmods/moborigins/entity/slime/OriginSlimeFollowOwnerGoal.java
@@ -24,7 +24,7 @@ public class OriginSlimeFollowOwnerGoal extends Goal {
 
     public OriginSlimeFollowOwnerGoal(OriginSlimeEntity tameable, double speed, float minDistance, float maxDistance, boolean leavesAllowed) {
         this.tameable = tameable;
-        this.world = tameable.world;
+        this.world = tameable.getWorld();
         this.speed = speed;
         this.navigation = tameable.getNavigation();
         this.minDistance = minDistance;