+++ src/main/java/me/ultrusmods/moborigins/mixin/AnimalEntityMixin.java
@@ -2,31 +2,44 @@ package me.ultrusmods.moborigins.mixin;
 
 import io.github.apace100.apoli.component.PowerHolderComponent;
 import me.ultrusmods.moborigins.power.ActionOnBreedAnimalPower;
+import net.minecraft.advancement.criterion.Criteria;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.passive.AnimalEntity;
 import net.minecraft.entity.passive.PassiveEntity;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.stat.Stats;
 import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import java.util.List;
+import java.util.Optional;
 
 @Mixin(AnimalEntity.class)
 public abstract class AnimalEntityMixin extends PassiveEntity {
+    @Shadow public abstract @Nullable ServerPlayerEntity getLovingPlayer();
+
     protected AnimalEntityMixin(EntityType<? extends PassiveEntity> entityType, World world) {
         super(entityType, world);
     }
 
-    @Inject(method = "breed", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/network/ServerPlayerEntity;incrementStat(Lnet/minecraft/util/Identifier;)V"), locals = LocalCapture.CAPTURE_FAILHARD)
-    public void breed$MobOrigins(ServerWorld world, AnimalEntity other, CallbackInfo ci, PassiveEntity passiveEntity, ServerPlayerEntity serverPlayerEntity) {
-        List<ActionOnBreedAnimalPower> powers = PowerHolderComponent.getPowers(serverPlayerEntity, ActionOnBreedAnimalPower.class).stream().filter(p -> p.shouldExecute(passiveEntity)).toList();
-        powers.forEach((actionOnBreedAnimal -> {
-            actionOnBreedAnimal.executeAction(passiveEntity);
-        }));
+    @Inject(
+        method = "breed(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/entity/passive/AnimalEntity;Lnet/minecraft/entity/passive/PassiveEntity;)V",
+        at = @At("HEAD")
+    )
+    public void breed$MobOrigins(ServerWorld world, AnimalEntity other, PassiveEntity child, CallbackInfo ci) {
+        Optional.ofNullable(this.getLovingPlayer()).or(() -> Optional.ofNullable(other.getLovingPlayer())).ifPresent(player -> {
+            List<ActionOnBreedAnimalPower> powers = PowerHolderComponent.getPowers(player, ActionOnBreedAnimalPower.class).stream().filter(p -> p.shouldExecute(child)).toList();
+            powers.forEach((actionOnBreedAnimal -> {
+                actionOnBreedAnimal.executeAction(child);
+            }));
+        });
+
     }
 }