+++ src/main/java/me/ultrusmods/moborigins/power/SnowTrailPower.java
@@ -35,8 +35,8 @@ public class SnowTrailPower extends Power {
             int j = MathHelper.floor(this.entity.getY());
             int k = MathHelper.floor(this.entity.getZ() + (double)((float)(l / 2 % 2 * 2 - 1) * 0.25F));
             BlockPos blockPos = new BlockPos(i, j, k);
-            if (this.entity.world.getBlockState(blockPos).isAir() && blockState.canPlaceAt(this.entity.getWorld(), blockPos)) {
-                this.entity.world.setBlockState(blockPos, blockState);
+            if (this.entity.getWorld().getBlockState(blockPos).isAir() && blockState.canPlaceAt(this.entity.getWorld(), blockPos)) {
+                this.entity.getWorld().setBlockState(blockPos, blockState);
             }
         }
     }